CPPCHECK Fix (additional)
- nullPointerRedundantCheck
- objectIndex

Also has additional corrections based on review comments.

+++ openair3/NAS/COMMON/EMM/MSG/AttachAccept.c
@@ -36,7 +36,7 @@ int decode_attach_accept(attach_accept_msg *attach_accept, uint8_t *buffer, uint
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ATTACH_ACCEPT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ATTACH_ACCEPT_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_u8_eps_attach_result(&attach_accept->epsattachresult, 0, *(buffer + decoded), len - decoded)) < 0)
@@ -205,7 +205,7 @@ int encode_attach_accept(attach_accept_msg *attach_accept, uint8_t *buffer, uint
   LOG_FUNC_IN;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ATTACH_ACCEPT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ATTACH_ACCEPT_MINIMUM_LENGTH, len);
 
   *(buffer + encoded) = (encode_u8_eps_attach_result(&attach_accept->epsattachresult) & 0x0f);
   encoded++;
