+++ openair3/NAS/COMMON/EMM/MSG/AttachReject.c
@@ -35,7 +35,7 @@ int decode_attach_reject(attach_reject_msg *attach_reject, uint8_t *buffer, uint
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ATTACH_REJECT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ATTACH_REJECT_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_emm_cause(&attach_reject->emmcause, 0, buffer + decoded, len - decoded)) < 0)
@@ -79,7 +79,7 @@ int encode_attach_reject(attach_reject_msg *attach_reject, uint8_t *buffer, uint
   int encode_result = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ATTACH_REJECT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ATTACH_REJECT_MINIMUM_LENGTH, len);
 
   if ((encode_result = encode_emm_cause(&attach_reject->emmcause, 0, buffer +
                                         encoded, len - encoded)) < 0)        //Return in case of error
