+++ openair3/NAS/COMMON/EMM/MSG/DetachAccept.c
@@ -33,7 +33,7 @@ int decode_detach_accept(detach_accept_msg *detach_accept, uint8_t *buffer, uint
 {
   uint32_t decoded = 0;
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, DETACH_ACCEPT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, DETACH_ACCEPT_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   return decoded;
@@ -43,7 +43,7 @@ int encode_detach_accept(detach_accept_msg *detach_accept, uint8_t *buffer, uint
 {
   int encoded = 0;
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, DETACH_ACCEPT_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, DETACH_ACCEPT_MINIMUM_LENGTH, len);
 
   return encoded;
 }
