+++ openair3/NAS/COMMON/EMM/MSG/GutiReallocationComplete.c
@@ -33,7 +33,7 @@ int decode_guti_reallocation_complete(guti_reallocation_complete_msg *guti_reall
 {
   uint32_t decoded = 0;
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, GUTI_REALLOCATION_COMPLETE_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, GUTI_REALLOCATION_COMPLETE_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   return decoded;
@@ -43,7 +43,7 @@ int encode_guti_reallocation_complete(guti_reallocation_complete_msg *guti_reall
 {
   int encoded = 0;
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, GUTI_REALLOCATION_COMPLETE_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, GUTI_REALLOCATION_COMPLETE_MINIMUM_LENGTH, len);
 
   return encoded;
 }
