+++ openair3/NAS/COMMON/EMM/MSG/TrackingAreaUpdateComplete.c
@@ -33,7 +33,7 @@ int decode_tracking_area_update_complete(tracking_area_update_complete_msg *trac
 {
   uint32_t decoded = 0;
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, TRACKING_AREA_UPDATE_COMPLETE_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, TRACKING_AREA_UPDATE_COMPLETE_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   return decoded;
@@ -43,7 +43,7 @@ int encode_tracking_area_update_complete(tracking_area_update_complete_msg *trac
 {
   int encoded = 0;
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, TRACKING_AREA_UPDATE_COMPLETE_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, TRACKING_AREA_UPDATE_COMPLETE_MINIMUM_LENGTH, len);
 
   return encoded;
 }
