+++ openair3/NAS/COMMON/ESM/MSG/EsmInformationRequest.c
@@ -34,7 +34,7 @@ int decode_esm_information_request(esm_information_request_msg *esm_information_
   uint32_t decoded = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ESM_INFORMATION_REQUEST_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ESM_INFORMATION_REQUEST_MINIMUM_LENGTH, len);
 
   /* Decoding mandatory fields */
   return decoded;
@@ -45,7 +45,7 @@ int encode_esm_information_request(esm_information_request_msg *esm_information_
   int encoded = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ESM_INFORMATION_REQUEST_MINIMUM_LENGTH, (int)len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ESM_INFORMATION_REQUEST_MINIMUM_LENGTH, len);
 
   return encoded;
 }
